chore: rework createComponent
