+++ packages/mp-runtime/src/computeChangedData.js
@@ -1,5 +1,33 @@
-import stringToPath from './stringToPath';
 import shallowEqual from './shallowEqual';
+
+const reLeadingDot = /^\./;
+// a..b
+// a[][]b
+const rePropName = /[^.[\]]+|\[(?:(-?\d+)|(["'])((?:(?!\2)[^\\]|\\.)*?)\2)\]|(?=(?:\.|\[\])(?:\.|\[\]|$))/g;
+/** Used to match backslashes in property paths. */
+const reEscapeChar = /\\(\\)?/g;
+const stringToPathCache = {};
+
+function stringToPath(stringPath) {
+  if (stringToPathCache[stringPath]) {
+    return stringToPathCache[stringPath];
+  }
+  const result = [];
+  if (reLeadingDot.test(stringPath)) {
+    result.push('');
+  }
+  stringPath.replace(rePropName, (match, num, quote, str) => {
+    let part = match;
+    if (quote) {
+      part = str.replace(reEscapeChar, '$1');
+    } else if (num) {
+      part = parseInt(num, 10);
+    }
+    result.push(part);
+  });
+  stringToPathCache[stringPath] = result;
+  return result;
+}
 
 function clone(obj, assumeArray) {
   if (!obj) {
