+++ packages/mp-runtime/src/index.js
@@ -1,8 +1,10 @@
 import createApp, { getApp } from './createApp';
 import createPage from './createPage';
+import createComponent from './createComponent';
 
 export {
   createApp,
   createPage,
+  createComponent,
   getApp,
 };
